[CMA-63] feat: 타이머 시작 후 버튼 보여지기 설정
각 버튼이 눌렸을 때 보이고 사라지는 버튼들을 주석으로 정리해둠 (헷갈림 방지..)
